requirements.txt
requests
python-dotenv
python-dateutil

pip install gspread oauth2client